Modified nuxt config and content-utils to better match framework specification
This commit makes a number of changes. One, content-utils is now a ES6 module to satisfy style guide. Next, routes property for generate property wraps the function from content-utils so it matches the spec (Note: a promise was insufficient, it has to be a **function** that returns a promise). Finally, we add a custom dynamic route for posts page as directory structure does not match generated routes