Fixed flash of white; added ability to store user preference for color scheme
This commit "fixes" flash of white experience on the first page load of the website. This is accomplished
by utilizing media queries instead of JS to set the default user experience. However, the tradeoff is
that for returning users of the site, if the toggle is used and permenant preference is specified,
this user now experiences the transition. This plus the added animation may not be ideal, but it is the tradeoff
being made"