Redid styles to fix breaking of default scroll behavior
This commit add a number of styles, naming by removing the flex container and adding manual breakpoints and styles
for sizing of menus and content. This is as the original solution, while convenient, effectively hijacked the window
scroll bar and there resulting in oddities such as the post being navigated to starting midway down the page.

The solution was to actually explicitly use position:fixed on the relevant elements and hardcording values as necessary
. As a sidenote, the webpage is more keyboard navigablable.