Change PostCard component to use full markdown formatting
This commit changes PostCard component into using the full markdown formatting instead of stripping it and presenting an excerpt.
This is to help faciliatate instances where rich media may be needed to be rendered such images and code. Otherwise, the author
would need to be mindful of the 140 word limit of the excerpt and avoid putting rich media in that area.

However, this is currently a tad naive in that it takes the entire markdwon and renders and the hiding is only achieved by CSS.
This may cause issues potentially with screen-readers and is a tad slow honestly. Ideally a way to either strip rich media or more
intelligent breaking system maybe a better option.