Added focus-visible styles; readded fonts.scss as global styles
This commit adds :focus-visible styles to remove outline on mouse navigation and improve
outline over default styles. This commit also adds focus-visible polyfill on unsupported
platforms.

Finally, this commit readds fonts stylesheet as global stylesheet. This is largely
a performance optimziation that reduces the number of times fonts are requested and
thereby reduce number of times relayout is done. The prior decision to remove was
a result of misunderstanding how @nuxt/style-resources works